Separate base dictionary_handler from root
