+++ speckle_connector/src/sketchup_model/dictionary/speckle_model_dictionary_handler.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'dictionary_handler'
+require_relative 'base_dictionary_handler'
 require_relative '../../constants/dict_constants'
 require_relative '../../constants/type_constants'
 
@@ -8,7 +8,7 @@ module SpeckleConnector
   module SketchupModel
     module Dictionary
       # Dictionary handler of the speckle model.
-      class SpeckleModelDictionaryHandler < DictionaryHandler
+      class SpeckleModelDictionaryHandler < BaseDictionaryHandler
         DICTIONARY_NAME = 'Speckle'
         # Writes initial data while speckle entity is creating first time.
         # @param sketchup_model [Sketchup::Model] Sketchup model to write data into it's attribute dictionary.
